feat: moveed mock transactions data to a separate file for better organization
